requirements.txt
Package Version
------- -------
pillow >= 11.1.0
